requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
paho-mqtt
